+++ test/unit/schema/typeGroupValidators/index.test.js
@@ -17,4 +17,71 @@ describe('Schema.typeGroupValidators', function () {
       'has the expected keys'
     )
   })
+
+  it('allows custom validation keywords', function () {
+    const STRING_OPS = JSData.Schema.STRING_OPS
+    const validationKeywords = JSData.Schema.validationKeywords
+    STRING_OPS.push('foo')
+    validationKeywords.foo = function (value, schema, opts) {
+      if (value !== 'bar') {
+        return [
+          {
+            actual: value,
+            expected: 'bar',
+            path: opts.path[0]
+          }
+        ]
+      }
+    }
+    const schema = new JSData.Schema({
+      type: 'object',
+      properties: {
+        thing: { foo: true, type: 'string', required: true },
+        name: { type: 'string' }
+      }
+    })
+    let errors = schema.validate({
+      name: 1234
+    })
+
+    assert.deepEqual(
+      errors,
+      [
+        {
+          expected: 'a value',
+          actual: 'undefined',
+          path: 'thing'
+        },
+        {
+          expected: 'one of (string)',
+          actual: 'number',
+          path: 'name'
+        }
+      ]
+    )
+
+    errors = schema.validate({
+      name: 'john',
+      thing: 'baz'
+    })
+    assert.deepEqual(
+      errors,
+      [
+        {
+          expected: 'bar',
+          actual: 'baz',
+          path: 'thing'
+        }
+      ]
+    )
+
+    errors = schema.validate({
+      name: 'john',
+      thing: 'bar'
+    })
+    assert.equal(errors, undefined)
+
+    delete validationKeywords.foo
+    STRING_OPS.pop()
+  })
 })